feat(characters): refactor character insertion to clear existing data and read from JSON
