+++ backend/src/repositories/characterRepository.ts
@@ -2,18 +2,13 @@ import * as fs from 'fs';
 import { AppDataSource } from '../libs/data-source';
 import { Character } from '../libs/entities/Character';
 
-export async function insertAllCharacters(): Promise<void> {
-  const jsonData = JSON.parse(fs.readFileSync('assets/characters/innitCharacters.json', 'utf-8'));
-  for (const character of jsonData.characters) {
-    const newCharacter = new Character();
-    newCharacter.characterId = character.characterId;
-    newCharacter.name = character.name;
-    newCharacter.speed = character.speed;
-    newCharacter.health = character.health;
-    newCharacter.dexterity = character.dexterity;
-    newCharacter.intelligence = character.intelligence;
-    await AppDataSource.getRepository(Character).save(newCharacter);
-  }
+async function deleteAll() {
+  await AppDataSource.getRepository(Character).clear();
+}
+
+export async function insertAllCharacters(chars: Character[]): Promise<void> {
+  await deleteAll();
+  await AppDataSource.getRepository(Character).save(chars);
 }
 export async function getAllCharactersRepo(): Promise<Character[]> {
   return await AppDataSource.getRepository(Character).find();
